requirements.txt
fastapi
python-dotenv
google-generativeai
python-multipart
uvicorn
pydantic
pydantic-settings